90% speed up and minor mode changes

+++ Sample_generation/generate_frame.m
@@ -1,5 +1,5 @@
 function frame = generate_frame(set)
-frame = zeros(set.mic.pixels_x,set.mic.pixels_y)';
+frame = int16(zeros(set.mic.pixels_x,set.mic.pixels_y)');
 end
 
 %Generates a matrix of zeroes with the dimensions of
